+++ pages/gallery/index.tsx
@@ -1,4 +1,4 @@
-import type { NextPage } from 'next'
+import { Suspense } from 'react'
 import Image from 'next/image'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
@@ -10,8 +10,9 @@ import cloudinary from '../../utils/cloudinary'
 import getBase64ImageUrl from '../../utils/generateBlurPlaceholder'
 import type { ImageProps } from '../../utils/types'
 import { useLastViewedPhoto } from '../../utils/useLastViewedPhoto'
+import { fetchImages } from '../../lib/data'
 
-const Gallery: NextPage = ({ images }: { images: ImageProps[] }) => {
+export default async function Gallery() {
   const router = useRouter()
   const { photoId } = router.query
   const [lastViewedPhoto, setLastViewedPhoto] = useLastViewedPhoto()
@@ -26,6 +27,8 @@ const Gallery: NextPage = ({ images }: { images: ImageProps[] }) => {
       setLastViewedPhoto(null)
     }
   }, [photoId, lastViewedPhoto, setLastViewedPhoto])
+
+  const images = await fetchImages();
   
   return (
     <>
@@ -45,31 +48,33 @@ const Gallery: NextPage = ({ images }: { images: ImageProps[] }) => {
           />
         )}
         <div className="columns-1 gap-4 sm:columns-2 xl:columns-3 2xl:columns-4 p-8">
-          {images.map(({ id, public_id, format, blurDataUrl }) => (
-            <Link
+          <Suspense>
+            {images.map(({ id, public_id, format }) => (
+              <Link
               key={id}
               href={`/gallery/?photoId=${id}`}
               as={`/gallery/p/${id}`}
               ref={id === Number(lastViewedPhoto) ? lastViewedPhotoRef : null}
               shallow
               className="after:content group relative mb-5 block w-full cursor-zoom-in after:pointer-events-none after:absolute after:inset-0 after:rounded-lg after:shadow-highlight"
-            >
-              <Image
-                alt="Mac photo"
-                className="transform rounded-lg brightness-90 transition will-change-auto group-hover:brightness-110"
-                style={{ transform: 'translate3d(0, 0, 0)' }}
-                placeholder="blur"
-                blurDataURL={blurDataUrl}
-                src={`https://res.cloudinary.com/${process.env.NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME}/image/upload/c_scale,w_720/${public_id}.${format}`}
-                width={720}
-                height={480}
-                sizes="(max-width: 640px) 100vw,
-                  (max-width: 1280px) 50vw,
-                  (max-width: 1536px) 33vw,
-                  25vw"
-              />
-            </Link>
-          ))}
+              >
+                <Image
+                  alt="Mac photo"
+                  className="transform rounded-lg brightness-90 transition will-change-auto group-hover:brightness-110"
+                  style={{ transform: 'translate3d(0, 0, 0)' }}
+                  placeholder="blur"
+                  src={`https://res.cloudinary.com/${process.env.NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME}/image/upload/c_scale,w_720/${public_id}.${format}`}
+                  width={720}
+                  height={480}
+                  sizes="(max-width: 640px) 100vw,
+                    (max-width: 1280px) 50vw,
+                    (max-width: 1536px) 33vw,
+                    25vw"
+                />
+              </Link>
+            ))}
+          </Suspense>
+
           {/* {test.map((element) => (
             <div>
               {element}
@@ -92,40 +97,39 @@ const Gallery: NextPage = ({ images }: { images: ImageProps[] }) => {
   )
 }
 
-export default Gallery
 
-export async function getStaticProps() {
-  const results = await cloudinary.v2.search
-    .expression(`folder:${process.env.CLOUDINARY_FOLDER}/*`)
-    .sort_by('public_id', 'asc')
-    .max_results(500)
-    .execute()
-  let reducedResults: ImageProps[] = []
+// export async function getStaticProps() {
+//   const results = await cloudinary.v2.search
+//     .expression(`folder:${process.env.CLOUDINARY_FOLDER}/*`)
+//     .sort_by('public_id', 'asc')
+//     .max_results(500)
+//     .execute()
+//   let reducedResults: ImageProps[] = []
 
-  let i = 0
-  for (let result of results.resources) {
-    reducedResults.push({
-      id: i,
-      height: result.height,
-      width: result.width,
-      public_id: result.public_id,
-      format: result.format,
-    })
-    i++
-  }
+//   let i = 0
+//   for (let result of results.resources) {
+//     reducedResults.push({
+//       id: i,
+//       height: result.height,
+//       width: result.width,
+//       public_id: result.public_id,
+//       format: result.format,
+//     })
+//     i++
+//   }
 
-  const blurImagePromises = results.resources.map((image: ImageProps) => {
-    return getBase64ImageUrl(image)
-  })
-  const imagesWithBlurDataUrls = await Promise.all(blurImagePromises)
+//   const blurImagePromises = results.resources.map((image: ImageProps) => {
+//     return getBase64ImageUrl(image)
+//   })
+//   const imagesWithBlurDataUrls = await Promise.all(blurImagePromises)
 
-  for (let i = 0; i < reducedResults.length; i++) {
-    reducedResults[i].blurDataUrl = imagesWithBlurDataUrls[i]
-  }
+//   for (let i = 0; i < reducedResults.length; i++) {
+//     reducedResults[i].blurDataUrl = imagesWithBlurDataUrls[i]
+//   }
 
-  return {
-    props: {
-      images: reducedResults,
-    },
-  }
-}
+//   return {
+//     props: {
+//       images: reducedResults,
+//     },
+//   }
+// }
